Using only one file to contain all the common functions required to run the app for both all the bills and 12 bills.
